Rename utility modules in test/integration
