+++ bokehjs/test/integration/layouts.ts
@@ -1,4 +1,4 @@
-import {display, fig, row, column, grid} from "./utils"
+import {display, fig, row, column, grid} from "./_util"
 
 import {Spacer, Tabs, Panel} from "@bokehjs/models/layouts"
 import {ToolbarBox} from "@bokehjs/models/tools/toolbar_box"
