+++ bokehjs/test/integration/plots.ts
@@ -1,4 +1,4 @@
-import {display, fig, row} from "./utils"
+import {display, fig, row} from "./_util"
 
 import {figure} from "@bokehjs/api/plotting"
 import {Location} from "@bokehjs/core/enums"
